Add Program Loop. Program logic removed from game_loop (renamed to gameplay)
