--- Desarrollo1_Pixeloids/src/engine/game_loop.h
@@ -1,11 +1,0 @@
-#ifndef GAME_LOOP
-#define GAME_LOOP
-
-namespace pixeloids_luchelli
-{
-
-void runGame();
-
-} // namespace pixeloids_luchelli
-
-#endif // GAME_LOOP
