+++ Desarrollo1_Pixeloids/src/engine/program_loop.h
@@ -1,0 +1,13 @@
+#ifndef PROGRAM_LOOP_H
+#define PROGRAM_LOOP_H
+
+namespace pixeloids_luchelli
+{
+
+	void runGame();
+
+}
+
+#endif // !PROGRAM_LOOP_H
+
+
