+++ Desarrollo1_Pixeloids/src/entities/asteroid.h
@@ -21,11 +21,8 @@ const int maxMediumAsteroids = 35;
 const int maxSmallAsteroids = 50;
 
 void InitializeAsteroid(Asteroid& asteroid, Vector2 position, Vector2 velocity, AsteroidSize size);
-
 void UpdateAsteroid(Asteroid& asteroid);
-
 void DrawAsteroid(Asteroid& asteroid, Texture2D& asteroidSprite);
-
 void DestroyAsteroid(Asteroid& asteroid, Asteroid asteroids[], int& asteroidCount, int& score, Sound asteroidDestroySfx);
 
 } // namespace pixeloids_luchelli
